python/requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
alembic==1.4.1
async-timeout==3.0.1
attrs==20.3.0
brotlipy==0.7.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
databricks-cli==0.14.3
docker==4.4.4
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Mako==1.1.4
MarkupSafe==1.1.1
mlflow==1.15.0
multidict==5.1.0
numpy==1.20.1
pandas==1.2.3
progress==1.5
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.15.6
psycopg2-binary==2.8.6
pyarrow==3.0.0
pycares==3.1.1
pycparser==2.20
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
smmap==4.0.0
SQLAlchemy==1.4.2
sqlparse==0.4.1
tabulate==0.8.9
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
yarl==1.6.3
